+++ crates/apis/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+* Added `random` feature to override `Math.random` implementation with one that sets the random seed on first use of `Math.random`
+
 ## 1.0.0 - 2023-05-17
 
 Initial release